fixed infinite render bug for DisplayTable, added txt file to store notes for oauth debugging
