+++ backend/google_response.txt
@@ -1,0 +1,3 @@
+http://localhost:4000/auth/google/callback?
+code=4%2F0AfgeXvsMNEPvMiVMT7B_SKdysPGPi0NNMvpEyv4SWF_epJRa6E9MTwZ6o_4Y6433utGr0g
+&scope=profile+https%3A%2F%2Fwww.googleapis.com%2Fauth%2Fuserinfo.profile
